+++ frontend/src/mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw';
 
-import { handlers } from '@mocks/handlers';
+import { handlers } from '@mocks/handlers/handlers';
 
 export const worker = setupWorker(...handlers);
