+++ frontend/src/mocks/handlers/detailStudyHandlers.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw';
 
-import { user } from '@mocks/memberHandlers';
+import { user } from '@mocks/handlers/memberHandlers';
 import studiesJSON from '@mocks/studies.json';
 
 import type { ApiStudy } from '@api/study';
